Fix-add-image+Delete
